Refactor the app and router section starts!
